refactor : PR 리뷰 반영 #7
- 관계자 가입 필요 여부 UserValidator로 분리
- UserAppender에서 모든 타입 별 가입 처리 담당
