+++ ea-domain/src/main/java/com/nbe2/domain/auth/AuthService.java
@@ -8,7 +8,6 @@ import org.springframework.transaction.annotation.Transactional;
 import lombok.RequiredArgsConstructor;
 
 import com.nbe2.domain.user.MedicalProfile;
-import com.nbe2.domain.user.MedicalUserAppender;
 import com.nbe2.domain.user.UserAppender;
 import com.nbe2.domain.user.UserProfile;
 import com.nbe2.domain.user.UserValidator;
@@ -19,17 +18,15 @@ public class AuthService {
 
     private final UserValidator userValidator;
     private final UserAppender userAppender;
-    private final MedicalUserAppender medicalUserAppender;
 
     @Transactional
     public void signUp(
-            UserProfile userProfile, Optional<Long> emergencyRoomId, Optional<Long> fileId) {
+            UserProfile userProfile, Optional<Long> emergencyRoomId, Optional<Long> licenseId) {
         userValidator.validate(userProfile.email());
 
-        if (emergencyRoomId.isPresent() || fileId.isPresent()) {
-            MedicalProfile medicalProfile =
-                    userValidator.validateMedicalProfile(emergencyRoomId, fileId);
-            medicalUserAppender.append(userProfile, medicalProfile);
+        if (userValidator.isMedicalUser(emergencyRoomId, licenseId)) {
+            MedicalProfile medicalProfile = userValidator.validate(emergencyRoomId, licenseId);
+            userAppender.append(userProfile, medicalProfile);
         } else {
             userAppender.append(userProfile);
         }
